Fix an off-by-one bug in get_handicap_rank_difference
Fix an off-by-one bug in the value of extra moves in
get_handicap_rank_difference, and rewrite the math using extra variables so
it's easier to visually verify.

The bug was in the calculation of the number of extra moves. The correct value is:

    num_extra_moves = handicap - 1 if handicap > 1 else 0

but the old code forgot the `- 1`.