Add --handicap-rank-difference-{small,19x19} options
Add options to compute the effective handicap rank difference, incorporating
board size, komi, and the point value of handicap stones.

This hits a math domain error when tallying because the head start for black is
sometimes enormous, such as 8 stones on a 9x9 board.

```
Processing OGS data
   2,739,321 /   30,811,698 games processed.  192.3s remainingTraceback (most recent call last):
  File "/Users/dexonsmith/Repos/online-go/goratings/analysis/analyze_glicko2_one_game_at_a_time.py", line 99, in <module>
    tally.add_glicko2_analytics(analytics)
  File "/Users/dexonsmith/Repos/online-go/goratings/analysis/util/TallyGameAnalytics.py", line 102, in add_glicko2_analytics
    self.prediction_cost[size][speed][rank][handicap] += - math.log(result.expected_win_rate if black_won else 1 - result.expected_win_rate)
                                                           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
ValueError: math domain error
```

Since the options are off-by-default, it'd fine to land as-is and then decide
what to do. Or we could fix first.

Here are the options:

- Update the tallying code to check for 0 before calling `math.log()`.
- Clamp effective rank between (e.g.) 40kyu and 15dan
- Stop rating games with a handicap rank difference bigger than (e.g.) 9
- Some combination above the above.