Add '--rating-history-db' to save ratings in a DB
Add `--rating-history-db` command-line option to save ratings in an sqlite3
database.

This adds a method to InMemoryStorage to save the rating history. It takes a
`category` string for future use with rating grids, where there's a different
InMemoryStorage for each rating category, but for now only `OneGameAtATime` has
been updated.

An aborted implementation used a new storage type OnDiskStorage instead of
InMemoryStorage, but it was way too slow to build the table incrementally.
Instead, this implementation writes the values in bulk at the end, using a
generator expression to visit the rating history.